+++ src/python/ensembl/genes/annotation-qc/metrics/config/software.json
@@ -1,0 +1,3 @@
+{
+    "agat": "/nfs/production/flicek/ensembl/shared_data/singularity_images/cache/agat_1.4.2--pl5321hdfd78af_0.sif"
+  }
